refactor(cli): split opportunity origin hints
